need to figure out talon methods

+++ src/com/nutrons/steamworks/Turret.java
@@ -13,7 +13,7 @@ public class Turret implements Subsystem {
   private final LoopSpeedController hoodMaster;
 
   private volatile double motorRotation;
-  private static final double MOTOR_ROTATIONS_TO_TURRET_ROTATIONS = 104 / 22;
+  private static final double MOTOR_ROTATIONS_TO_TURRET_ROTATIONS = (double) 104 / 22;
 
   public Turret(Flowable<Double> angle, LoopSpeedController master) {
     this.angle = angle;
